feat(cli): package entrypoint jva; add MIT LICENSE; convert to setuptools; keep run.py launcher
