+++ src/jva/cli.py
@@ -1,0 +1,5 @@
+"""Console entry for `jva` command: delegates to jva.run.main"""
+
+from .run import main
+
+__all__ = ["main"]
